pyproject.toml
[project]
name = "secure-enterprise-rag"
version = "0.1.0"
description = "Self-hosted RAG solution ensuring data privacy. Built with Open Web UI, Milvus, and Mistral."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.124.0",
    "grpcio>=1.76.0",
    "httpx>=0.28.1",
    "llama-index-core>=0.14.10",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-llms-openai-like>=0.5.3",
    "llama-index-vector-stores-milvus>=0.9.4",
    "openai>=2.9.0",
    "pydantic-settings>=2.12.0",
    "pymilvus>=2.6.5",
    "pypdf>=6.4.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.1",
    "unstructured[pdf]>=0.18.21",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.8",
]
